requirements.txt
astroid==2.6.5
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.21.2
psycopg2-binary==2.9.1
pylint==2.9.5
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
tabulate==0.8.9
toml==0.10.2
wrapt==1.12.1
